+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/plugins/project/IntelliJPlatformBasePlugin.kt
@@ -7,10 +7,14 @@ import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.api.plugins.JavaLibraryPlugin
 import org.gradle.api.plugins.JavaPluginExtension
+import org.gradle.api.tasks.compile.JavaCompile
 import org.gradle.jvm.toolchain.JavaLanguageVersion
+import org.gradle.jvm.toolchain.JavaToolchainService
 import org.gradle.kotlin.dsl.apply
-import org.gradle.kotlin.dsl.configure
+import org.gradle.kotlin.dsl.assign
 import org.gradle.kotlin.dsl.get
+import org.gradle.kotlin.dsl.getByType
+import org.gradle.kotlin.dsl.withType
 import org.gradle.plugins.ide.idea.IdeaPlugin
 import org.gradle.plugins.ide.idea.model.IdeaModel
 import org.jetbrains.intellij.platform.gradle.Constants.CACHE_DIRECTORY
@@ -38,6 +42,9 @@ import org.jetbrains.intellij.platform.gradle.services.registerClassLoaderScoped
 import org.jetbrains.intellij.platform.gradle.tasks.*
 import org.jetbrains.intellij.platform.gradle.tasks.aware.*
 import org.jetbrains.intellij.platform.gradle.utils.*
+import org.jetbrains.kotlin.gradle.dsl.JvmTarget
+import org.jetbrains.kotlin.gradle.tasks.KotlinJvmCompile
+import kotlin.getValue
 
 abstract class IntelliJPlatformBasePlugin : Plugin<Project> {
 
@@ -119,7 +126,7 @@ abstract class IntelliJPlatformBasePlugin : Plugin<Project> {
                         $message
                         Please ensure there is a single IntelliJ Platform dependency defined in your project and that the necessary repositories, where it can be located, are added.
                         See: https://plugins.jetbrains.com/docs/intellij/tools-intellij-platform-gradle-plugin-dependencies-extension.html
-                        """.trimIndent()
+                        """.trimIndent(),
                     )
                 }
 
@@ -219,7 +226,7 @@ abstract class IntelliJPlatformBasePlugin : Plugin<Project> {
                     addAllLater(
                         project.provider {
                             dependenciesHelper.createJetBrainsRuntimeObtainedDependency()
-                        }
+                        },
                     )
                 }
             }
@@ -284,8 +291,8 @@ abstract class IntelliJPlatformBasePlugin : Plugin<Project> {
                         project.providers[GradleProperties.VerifyPluginDefaultRecommendedIdes].flatMap { enabled ->
                             when {
                                 enabled &&
-                                    intellijPluginVerifierIdesDependencyConfiguration.dependencies.isEmpty() &&
-                                    intellijPluginVerifierIdesLocalConfiguration.dependencies.isEmpty() ->
+                                        intellijPluginVerifierIdesDependencyConfiguration.dependencies.isEmpty() &&
+                                        intellijPluginVerifierIdesLocalConfiguration.dependencies.isEmpty() ->
                                     dependenciesHelper.createIntelliJPluginVerifierIdeDependencies(
                                         dependenciesHelper.createRecommendedPluginVerifierIdesValueSource(),
                                     )
@@ -325,7 +332,8 @@ abstract class IntelliJPlatformBasePlugin : Plugin<Project> {
                 description = "Java Compiler used by Ant tasks",
             ) {
                 defaultDependencies {
-                    val addDefaultDependenciesProvider = project.providers[GradleProperties.AddDefaultIntelliJPlatformDependencies]
+                    val addDefaultDependenciesProvider =
+                        project.providers[GradleProperties.AddDefaultIntelliJPlatformDependencies]
                     val instrumentCodeProvider = project.extensionProvider.flatMap { it.instrumentCode }
 
                     addAllLater(
@@ -348,7 +356,7 @@ abstract class IntelliJPlatformBasePlugin : Plugin<Project> {
 
             val intellijPlatformTestDependenciesConfiguration = create(
                 name = Configurations.INTELLIJ_PLATFORM_TEST_DEPENDENCIES,
-                description = "IntelliJ Platform Test Dependencies"
+                description = "IntelliJ Platform Test Dependencies",
             ) {
                 extendsFrom(
                     intellijPlatformTestPluginConfiguration,
@@ -358,13 +366,13 @@ abstract class IntelliJPlatformBasePlugin : Plugin<Project> {
             }
             create(
                 name = Configurations.INTELLIJ_PLATFORM_CLASSPATH,
-                description = "IntelliJ Platform Classpath resolvable configuration"
+                description = "IntelliJ Platform Classpath resolvable configuration",
             ) {
                 extendsFrom(intellijPlatformConfiguration)
             }
             create(
                 name = Configurations.INTELLIJ_PLATFORM_TEST_CLASSPATH,
-                description = "IntelliJ Platform Test Classpath resolvable configuration"
+                description = "IntelliJ Platform Test Classpath resolvable configuration",
             ) {
                 extendsFrom(
                     intellijPlatformConfiguration,
@@ -374,7 +382,7 @@ abstract class IntelliJPlatformBasePlugin : Plugin<Project> {
             }
             create(
                 name = Configurations.INTELLIJ_PLATFORM_TEST_RUNTIME_CLASSPATH,
-                description = "IntelliJ Platform Test Runtime Classpath resolvable configuration"
+                description = "IntelliJ Platform Test Runtime Classpath resolvable configuration",
             ) {
                 attributes {
                     attributes.attribute(Attributes.kotlinJPlatformType, "jvm")
@@ -386,7 +394,7 @@ abstract class IntelliJPlatformBasePlugin : Plugin<Project> {
             }
             create(
                 name = Configurations.INTELLIJ_PLATFORM_TEST_RUNTIME_FIX_CLASSPATH,
-                description = "IntelliJ Platform Test Runtime Fix Classpath"
+                description = "IntelliJ Platform Test Runtime Fix Classpath",
             ) {
                 defaultDependencies {
                     addAllLater(
@@ -405,7 +413,7 @@ abstract class IntelliJPlatformBasePlugin : Plugin<Project> {
             }
             create(
                 name = Configurations.INTELLIJ_PLATFORM_RUNTIME_CLASSPATH,
-                description = "IntelliJ Platform Runtime Classpath resolvable configuration"
+                description = "IntelliJ Platform Runtime Classpath resolvable configuration",
             ) {
                 attributes {
                     attributes.attribute(Attributes.kotlinJPlatformType, "jvm")
@@ -486,19 +494,43 @@ abstract class IntelliJPlatformBasePlugin : Plugin<Project> {
             }
         }
 
-        project.extensions.configure<JavaPluginExtension> {
-            toolchain.languageVersion.convention(
-                project.cachedProvider {
-                    project.configurations[Configurations.INTELLIJ_PLATFORM_DEPENDENCY]
-                        .asLenient
-                        .productInfo()
-                        .buildNumber
-                        .toVersion()
-                        .toPlatformJavaVersion()
-                        .majorVersion
-                        .toInt()
-                }.map(JavaLanguageVersion::of)
+        // Setup a default Java language based on the IntelliJ Platform dependency
+        val intellijPlatformJavaLanguageVersion = project.cachedProvider {
+            project.configurations[Configurations.INTELLIJ_PLATFORM_DEPENDENCY]
+                .asLenient
+                .productInfo()
+                .buildNumber
+                .toVersion()
+                .toPlatformJavaVersion()
+                .majorVersion
+                .toInt()
+        }.map(JavaLanguageVersion::of)
+        val javaExtension = project.extensions.getByType<JavaPluginExtension>()
+        val requestedJavaLanguageVersion = javaExtension.toolchain.languageVersion.orElse(intellijPlatformJavaLanguageVersion)
+        val javaToolchainService = project.extensions.getByType<JavaToolchainService>()
+
+        project.tasks.withType<JavaCompile>().configureEach {
+            options.release.convention(requestedJavaLanguageVersion.map { it.toString().toInt() })
+            javaCompiler.convention(
+                javaToolchainService.compilerFor {
+                    languageVersion = requestedJavaLanguageVersion
+                    vendor = javaExtension.toolchain.vendor
+                    implementation = javaExtension.toolchain.implementation
+                }
             )
+        }
+
+        project.pluginManager.withPlugin(Plugins.External.KOTLIN) {
+            val javaLauncher = javaToolchainService.launcherFor {
+                languageVersion = requestedJavaLanguageVersion
+                vendor = javaExtension.toolchain.vendor
+                implementation = javaExtension.toolchain.implementation
+            }
+
+            project.tasks.withType<KotlinJvmCompile>().configureEach {
+                compilerOptions.jvmTarget.convention(requestedJavaLanguageVersion.map { JvmTarget.fromTarget(it.toString()) })
+                kotlinJavaToolchain.toolchain.use(javaLauncher)
+            }
         }
 
         IntelliJPlatformExtension.register(project, target = project).let { intelliJPlatform ->